requirements.txt
absl-py==1.2.0
addict==2.4.0
aiohttp==3.8.3
aiosignal==1.3.1
albumentations==1.3.0
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-timeout==4.0.2
attrs==22.1.0
av==10.0.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
certifi==2023.5.7
charset-normalizer==2.1.1
contextlib2==21.6.0
contourpy==1.1.0
cycler==0.11.0
debugpy @ file:///tmp/build/80754af9/debugpy_1637091799509/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
depix @ file:///data/lqk/Depix-main
einops==0.4.1
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
filelock==3.12.0
fire==0.4.0
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2022.11.0
ftfy==6.1.1
h5py==3.7.0
huggingface-hub==0.14.1
idna==3.3
imageio==2.28.1
importlib-resources==5.12.0
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1655369107642/work
ipython @ file:///opt/conda/conda-bld/ipython_1657652213665/work
ipywidgets @ file:///home/conda/feedstock_root/build_artifacts/ipywidgets_1680023138361/work
jedi @ file:///tmp/build/80754af9/jedi_1644297102865/work
joblib==1.2.0
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1633454794268/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1669775081219/work
jupyterlab-widgets @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_widgets_1684239754370/work
kiwisolver==1.4.4
lazy_loader==0.2
lightning-utilities==0.3.0
littleutils==0.2.2
matplotlib==3.7.1
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
MedPy==0.4.0
ml-collections==0.1.1
mmcv==1.7.0
multidict==6.0.3
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1664684991461/work
networkx==3.1
numpy==1.23.3
ogb==1.3.6
open-clip-torch==2.19.0
opencv-python==4.6.0.66
opencv-python-headless==4.7.0.72
outdated==0.2.2
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1681337016113/work
pandas==1.5.2
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.2.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
protobuf==3.20.1
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptflops==0.6.9
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pyparsing==3.0.9
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytorch-lightning==1.8.3.post1
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==19.0.2
qudida==0.0.4
regex==2023.5.5
requests==2.28.1
safetensors==0.3.1
scikit-fuzzy==0.4.2
scikit-image==0.20.0
scikit-learn==1.2.2
scikit-video==1.1.11
scipy==1.9.1
sentencepiece==0.1.99
SimpleITK==2.2.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0.post5
some-package==0.1
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
tensorboardX==2.5.1
termcolor==2.1.1
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.8.23.dev0
torch==1.12.1+cu116
torchaudio==0.12.1+cu116
torchmetrics==0.11.0
torchstat==0.0.7
torchvision==0.13.1+cu116
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1648827245914/work
tqdm==4.64.1
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
widgetsnbextension @ file:///home/conda/feedstock_root/build_artifacts/widgetsnbextension_1680021576815/work
wilds==2.0.0
yacs==0.1.8
yapf==0.32.0
yarl==1.8.2
zipp==3.15.0
